split platform dependent classes in separate files
